+++ src/psffontop.h
@@ -20,7 +20,7 @@ struct unicode_list {
 	struct unicode_seq *seq;
 };
 
-extern int readpsffont(FILE *fontf, char **allbufp, int *allszp,
+extern int readpsffont(FILE *fontf, char **allbufp, size_t *allszp,
                        char **fontbufp, int *fontszp,
                        int *fontwidthp, int *fontlenp, int fontpos0,
                        struct unicode_list **uclistheadsp);
